+++ distribution/win11/EggR.Win11Distribution.psm1
@@ -1,7 +1,7 @@
 Set-StrictMode -Version Latest
 $ErrorActionPreference = 'Stop'
 
-$script:ExpectedExtensionId = 'integratedpower.integrated-power'
+$script:ExpectedExtensionId = 'EggR.integrated-power'
 $script:AllowedKnowledgeTargets = @(
     'eggr-roots.ps1',
     'set-eggr-roots.ps1',